test(mongo): refactored in memory mongo db to start only once per JVM
Starting and stopping the in memory database inbetween tests causes odd connection isues on mac osx and broken pipe errors
on linux. Refactored the test code so that the start of the database is done once and each test cleans up the data
collections between scenarios.
